refactored components fix
