+++ src/components/AdminTable/AdminTableModal/FilteringContent/index.ts
@@ -1,0 +1,3 @@
+import FilteringContent from "./FilteringContent";
+
+export default FilteringContent;
