+++ src/components/AdminTable/AdminTableModal/SortingContent/index.ts
@@ -1,0 +1,3 @@
+import SortingContent from "./SortingContent";
+
+export default SortingContent;
